baseline

+++ trainer.py
@@ -776,7 +776,7 @@ class UnsupervisedTrainer():
     
     def train(self):
         
-        tens = torch.ones(self.criterion.m.shape[0], dtype=metadata.dtype)*10
+        tens = torch.ones(self.criterion.m.shape[0], dtype=torch.long)*10
         if torch.cuda.is_available() and self.configer.cuda: tens = tens.cuda()
         self.writer.add_embedding(self.criterion.m, tens, global_step=self.cur_epoch)
